+++ include/gtirb_pprinter/MasmPrettyPrinter.hpp
@@ -159,6 +159,10 @@ protected:
   void printString(std::ostream& Stream, const gtirb::DataBlock& Block,
                    uint64_t Offset, bool NullTerminated = true) override;
 
+  std::optional<uint64_t>
+  adjustAlignment(const gtirb::DataBlock& Block,
+                  const std::optional<uint64_t>& Align) override;
+
 private:
   gtirb::Addr BaseAddress;
   const gtirb::Symbol* ImageBase;