+++ tests/gtirb_helpers.py
@@ -105,6 +105,9 @@ def add_standard_aux_data_tables(m: gtirb.Module) -> None:
         )
 
     elif m.file_format == gtirb.Module.FileFormat.PE:
+        m.aux_data["alignment"] = gtirb.AuxData(
+            type_name="mapping<UUID,uint64_t>", data=dict()
+        )
         m.aux_data["peExportEntries"] = gtirb.AuxData(
             type_name="sequence<tuple<uint64_t,int64_t,string>>", data=list()
         )